--- tests/e2e/jupyverse_adapter/types_.py
@@ -1,5 +1,0 @@
-from typing import TypeAlias
-
-
-KernelId: TypeAlias = str
-SessionId: TypeAlias = str
